move nvme-zfs func out of types into own package to avoid dependency pollution
